replace refreshTokenResultReceiver with local variable.
